NFC: Modify third party libnfc-nci
The third party libnfc-nci has to coexist with the AOSP baseline.
Furthermore, it has to be compatible with the NQx chip family.

Modified makefiles and config files to reflect NQx naming scheme.
Modified HAL to avoid modifications to libhardware for new ioctl
function pointer.

+++ halimpl/pn54x/tml/phTmlNfc_i2c.c
@@ -362,7 +362,7 @@ NFCSTATUS phTmlNfc_set_pid(void *pDevHandle, long pid)
         return NFCSTATUS_FAILED;
     }
 
-    ret = ioctl((int32_t)pDevHandle, P544_SET_NFC_SERVICE_PID, pid);
+    ret = ioctl((intptr_t)pDevHandle, P544_SET_NFC_SERVICE_PID, pid);
     return ret;
 }
